"Hello, world!" with React
Note: Don't want to `require('react')` as we're doing at the top of
the file. Prefer to just do the granular `require`s to specific parts
of React that we need. However, removing this line causes an error
in React's code, so keep this `require` for now with hopes to remove it
later.