Remove call to `guesses.onCompleted()`
Can't figure out why this is needed, and removing it doesn't seem to
break anything.

Not sure why I added this before. Why would we want to end the
`guesses` sequence when there's a correct guess? Suspect I
included it at a moment when I thought that we'd have a new `guesses`
stream for every new note that's displayed, when we ended up just
having a single `guesses` stream.

I suspect that